+++ wasm/src/base/export.rs
@@ -1,0 +1,12 @@
+use crate::base::indices::{GlobalIndex, FunctionIndex, MemoryIndex};
+
+pub struct Export {
+    pub name: String,
+    pub export_description: ExportDescription,
+}
+
+pub enum ExportDescription {
+    Function(FunctionIndex),
+    Memory(MemoryIndex),
+    Global(GlobalIndex),
+}
