+++ wasm/src/base/memory.rs
@@ -1,0 +1,5 @@
+// ==Limits==
+pub enum Limit {
+    MinToInfinity { min: u32 },
+    MinMax { min: u32, max: u32 },
+}
