+++ wasm/src/base/mod.rs
@@ -1,0 +1,5 @@
+pub mod export;
+pub mod import;
+pub mod indices;
+pub mod memory;
+pub mod types;
