+++ wasm/src/binary_format/mod.rs
@@ -1,0 +1,8 @@
+pub mod primitives;
+pub mod sections;
+
+mod export;
+mod import;
+mod indices;
+mod memory;
+mod types;
