+++ wasm/src/binary_format/primitives/byte_stream.rs
