+++ wasm/src/binary_format/primitives/mod.rs
@@ -1,0 +1,2 @@
+pub mod byte_stream;
+pub mod encoder;
